This is a model of the project.
It contains pictures of the model as well as a video showing the working. In the working, the fingers represent cars (passing by the IR Sensor) on a road. When a road has some non-zero traffic density while the other 3 roads have zero traffic density, the road with non-zero traffic density gets a green light. In general, the road with highest traffic density is given the highest priority and its light becomes green.